+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Bug-fixes within the same version aren't needed
 ## Master
 
 * change path of jest executable in order to support running jest in yarn workspaces - @glentakahashi
+* refactor editor decorations and load icons from `vscode-codicons` [@Tymek](https://github.com/Tymek)
+* move test status indicator from editor text to gutter [@Tymek](https://github.com/Tymek)
 
 -->